Temporarily disable check for ubuntu-greeter
This is a security measure where the mtp-server is not started if the
phone is locked, as that someone can potentially plug in your phone and
then access it.

However, the code explicitly checks for the ubuntu-greeter which makes
it break in non ubuntu environments. Untill we find a proper cross
desktop way of figuring out when the phone is unlocked, I'm temporarily
disabling this feature and always starting the mtp-server.

This patch should be reverted in the future.